+++ CHANGELOG.md
@@ -26,3 +26,5 @@ All notable changes to this project will be documented in this file.
 - More tiled patterns.
 - Pattern viewer.
 - util.lua.
+- Option for right aligned menu UI.
+- Stories and Areas sub menu in home screen.